requirements.txt
# ═══════════════════════════════════════════════════════════════
# Jarvis OS — Python Dependencies
# ═══════════════════════════════════════════════════════════════

# Core
aiohttp>=3.9,<4.0
pyyaml>=6.0,<7.0
httpx>=0.27,<1.0

# LLM Providers (install what you need)
openai>=1.30,<2.0
anthropic>=0.30,<1.0

# Memory
chromadb>=0.4.24,<0.5.0

# Tools
duckduckgo-search>=6.0,<7.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<6.0
